refactor: Google Credential 관련 클래스 구조 변경
- GoogleCredentialHandler와 GoogleCredentialRequestManager를 GoogleCredentialManager로 통합
- AuthRepository 및 AuthDefaultRepository 추가
- LoginResult 모델 추가
