+++ android/app/src/main/java/com/yagubogu/domain/repository/AuthRepository.kt
@@ -1,0 +1,9 @@
+package com.yagubogu.domain.repository
+
+import com.yagubogu.domain.model.LoginResult
+
+interface AuthRepository {
+    suspend fun signInWithGoogle(): LoginResult
+
+    suspend fun signOutWithGoogle(): Result<Boolean>
+}
